Update Game.swift
Here's one more problem. When rook did move NOT from file == 0, then it was always counted as king-side-castle excluding:

        if piece.kind == .rook {
            let kind: PieceKind = move.from.file == 0 ? .queen : .king

So probably a solution can be just checking if move is from or to one of the corner squares, and excluding that castling option. In fact it doesn't matter which piece moves in this case, this can be both capturing rook or rook move, or even just any piece move to/from corner - we have to exclude corner square for all cases.
And this probably requires extra tests to be sure that it works correct.